+++ resources/views/card/show.blade.php
@@ -1,4 +1,5 @@
 <p>
     {{ $card->name }}
     {{ $card->description }}
+    {{ $card->type_id}}
 </p>