+++ resources/views/welcome.blade.php
@@ -19,8 +19,13 @@
     <div class="justify-center align-middle flex ">
         <header>
                         @if (Route::has('login'))
+
                             <nav class="-mx-3 flex flex-1 justify-end ">
                                 @auth
+                                <form action="{{ route('logout') }}" method="POST">
+                                    @csrf
+                                    <button type="submit">Logout</button>
+                                </form>
                                     <a
                                         href="{{ url('/dashboard') }}"
                                         class="rounded-md px-3 py-2 text-black ring-1 ring-transparent transition hover:text-black/70 focus:outline-none focus-visible:ring-[#FF2D20] dark:text-white dark:hover:text-white/80 dark:focus-visible:ring-white"