+++ routes/web.php
@@ -13,14 +13,14 @@ Route::get('/home', function () {
 
 !!  // route aanmaken met gebruik van controller
 
+//// passeer parameters via de url
+//Route::get('products/{id}', function(string $id){
+//    return view('product-id', [
+//        'id'=>$id,
+//    ]);
+//});
 
-// passeer parameters via de url
-Route::get('products/{id}', function(string $id){
-    return view('product-id', [
-        'id'=>$id,
-    ]);
-});
-
+// General related routes
 Route::get('/dashboard', function () {
     return view('dashboard');
 })->middleware(['auth', 'verified'])->name('dashboard');
@@ -28,17 +28,27 @@ Route::get('/dashboard', function () {
 Route::get('/contact', function () {
     return view('contact');
 })->name('contact');
+// End general related routes
 
+// Card related routes
+Route::get('/cards',[CardController::class, 'index'])->name('cards.index');
+Route::resource('/cards', CardController::class);
+// End card related routes
+
+// Types related routes
+Route::resource('/types', TypeController::class);
+// End related routes
+
+// Everything in the 'auth group' needs authentication to access.
 Route::middleware('auth')->group(function () {
+
+    //Profile related
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-    Route::post('/cards', [CardController::class, 'store']);
+
+    //Cards related
+    Route::post('/cards', [CardController::class, 'store'])->name('index.create');
 });
-
-Route::resource('/cards', CardController::class);
-
-Route::resource('/types', TypeController::class);
-
 
 require __DIR__.'/auth.php';